// src/app/models/config.model.ts
export class Config {

  //Contracts
  wbnbContract?: string;
  usdtContract?: string;

  //SwapRouter
  quickSwapRouter?: string;
  pancakeSwapRouter?: string;
  cofeeSwapRouter?: string;
  router?: string;

  //SwapFactory
  quickSwapFactory?: string;
  pancakeSwapFactory?: string;
  cofeeSwapFactory?: string;
  factory?: string;

  bscMainNetURL?: string;
}